requirements.txt
absl-py==0.8.1
aes==1.0.0
aiofiles==0.4.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.0
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
crypto==1.4.1
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
docker==4.1.0
factories==1.2.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h11==0.8.1
h2==3.1.1
h5py==2.10.0
hpack==3.0.0
httpcore==0.3.0
httplib2==0.9.2
httptools==0.0.13
hyperframe==5.2.0
idna==2.6
imutils==0.5.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpad==1.0.0
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.1.1
MarkupSafe==1.0
multidict==4.5.2
Naked==0.1.31
netifaces==0.10.4
numpy==1.17.3
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.1.26
opt-einsum==3.1.0
pexpect==4.2.1
Pillow==6.2.1
protobuf==3.10.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
pymongo==3.8.0
PyNaCl==1.1.2
pyRFC3339==1.0
python-apt==1.6.4
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzbar==0.1.8
qute==1.0.10
reportlab==3.4.0
requests==2.18.4
requests-async==0.5.0
requests-unixsocket==0.1.5
rfc3986==1.3.2
sanic==19.3.1
Sanic-Cors==0.9.8.post1
sanic-openapi==0.5.3
Sanic-Plugins-Framework==0.8.2
scribble==1.0.0
SecretStorage==2.3.1
shellescape==3.4.1
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.1
termcolor==1.1.0
ubuntu-drivers-common==0.0.0
ufw==0.36
ujson==1.35
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
uvloop==0.13.0
virtualenv==16.7.6
wadllib==1.3.2
websocket-client==0.56.0
websockets==6.0
Werkzeug==0.16.0
wrapt==1.11.2
xkit==0.0.0
zope.interface==4.3.2
